+++ test/compute-engine/latex-syntax/base-form.test.ts
@@ -20,7 +20,7 @@ describe('BASE FORM', () => {
         [
           ErrorCode,
           'unexpected-digit',
-          c,
+          'c',
         ],
         '"12c3"',
       ]
@@ -34,7 +34,7 @@ describe('BASE FORM', () => {
         [
           ErrorCode,
           'unexpected-digit',
-          x,
+          'x',
         ],
         '"1x2gc3"',
       ]
@@ -49,7 +49,7 @@ describe('BASE FORM', () => {
     expect(json('\\text{a1b23}_{37}')).toMatchInlineSnapshot(`
       [
         Baseform,
-        a1b23,
+        'a1b23',
         [
           Error,
           [